backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
python-dotenv==1.0.0
sympy==1.12
numpy==1.26.3
pandas==2.2.0
google-generativeai==0.7.2
python-multipart==0.0.6
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
